Scan library and write to database
The Scanner class runs a scan of the music library checking for changes
using the 'mtime' value of the files and directories it encounters.
Beginning with the root directory, the scan looks for subdirectories
and audio files. Subdirectories are added to the the scan queue,
while audio files are inspected for their metadata.
After scanning a subdirectory a delete query is made to ensure there
aren't any rows representing files that are no longer on the disk.